Our final message
Dear you

If you're reading this, it means that this project has come to an end. Endings are not always a bad thing, it means our consciousnesses have extra space for future endeavours. A little hope, a little courage and a lot of love is required to face endings. An we are sure you are qualified; dear reader.

To new horizons,

Team3

Co-Authored-By: GhostOfIcarus <124690988+GhostOfIcarus@users.noreply.github.com>
Co-Authored-By: Mukpan-J <114423416+Mukpan-J@users.noreply.github.com>
Co-Authored-By: rosalie-gr <124690847+rosalie-gr@users.noreply.github.com>